# backend/src/main/java/com/logiconnect/platform/notification/dto/NotificationResponse.java
package com.logiconnect.platform.notification.dto;

import com.logiconnect.platform.notification.entity.Notification;
import com.logiconnect.platform.notification.entity.NotificationType;

import java.time.Instant;
import java.util.UUID;

public record NotificationResponse(
        UUID id,
        NotificationType type,
        String title,
        String message,
        String referenceType,
        UUID referenceId,
        boolean isRead,
        Instant readAt,
        Instant createdAt
) {
    public static NotificationResponse fromEntity(Notification notification) {
        if (notification == null) return null;
        return new NotificationResponse(
                notification.getId(),
                notification.getType(),
                notification.getTitle(),
                notification.getMessage(),
                notification.getReferenceType(),
                notification.getReferenceId(),
                notification.isRead(),
                notification.getReadAt(),
                notification.getCreatedAt()
        );
    }
}