# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LogiConnect Enterprise REST API
  description: |
    Enterprise internal communication and collaboration platform for logistics operations.
    Supports organizational hierarchy, granular RBAC, secure messaging, company broadcasts,
    document management, meeting coordination, and compliance audit logs.
  version: 1.0.0
  contact:
    name: LogiConnect Architecture Team
    email: architecture@logiconnect.internal
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
  - url: https://api.logiconnect.internal/api/v1
    description: Corporate Production Gateway
...
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List User Notifications
      description: Returns a paginated list of notifications for the authenticated user, ordered newest first.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
        - name: isRead
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Notification feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPageEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get Unread Notification Count
      description: Returns the total unread notification count for the authenticated employee badge.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Unread notification count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/{id}/read:
    post:
      tags: [Notifications]
      summary: Mark Notification Read
      description: Marks a single notification as read. Scoped strictly to the current user (IDOR safe; returns 404 if notification belongs to another user).
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications/read-all:
    post:
      tags: [Notifications]
      summary: Mark All Notifications Read
      description: Marks all unread notifications for the currently authenticated employee as read in a single batch.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationBatchReadEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==========================================================================
  # 13. SEARCH API
  # ==========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageQuery:
      name: page
      in: query
      description: 0-indexed page number
      schema:
        type: integer
        default: 0
    SizeQuery:
      name: size
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        default: 20
        maximum: 100
...
  responses:
...
    NotFound:
      description: Target Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_FAILED"
            message:
              type: string
              example: "Provided payload violates validation constraints"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  rejectedValue:
                    type: string
                  issue:
                    type: string
        timestamp:
          type: string
          format: date-time
          example: "2026-08-31T12:00:00Z"
        path:
          type: string
          example: "/api/v1/employees"
...
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - MESSAGE
            - GROUP_MESSAGE
            - CHANNEL_MESSAGE
            - ANNOUNCEMENT
            - URGENT_ANNOUNCEMENT
            - ACKNOWLEDGEMENT_REQUIRED
            - MEETING_INVITATION
            - MEETING_UPDATED
            - MEETING_CANCELLED
            - MENTION
            - SECURITY
        title:
          type: string
        message:
          type: string
        referenceType:
          type: string
        referenceId:
          type: string
          format: uuid
        isRead:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    NotificationPageEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/NotificationResponse'
            page:
              type: integer
            size:
              type: integer
            totalElements:
              type: integer
            totalPages:
              type: integer
            first:
              type: boolean
            last:
              type: boolean
        message:
          type: string
    NotificationEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/NotificationResponse'
        message:
          type: string
    UnreadCountEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            unreadCount:
              type: integer
              format: int64
        message:
          type: string
    NotificationBatchReadEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            markedCount:
              type: integer
        message:
          type: string